OCT_App/requirements.txt
streamlit
torch
torchvision
opencv-python-headless
pytorch-grad-cam
numpy
Pillow